V0.2.0/jun/sep model api 모델 기능 나누기 (#109)
 모델과 기능별로 나누고 totalapi에서 부르기
